Add ListSDKS method to ProcessOutput class and verbose
Updated package versions and added new features

Updated the versions of `RSCG_WhatIAmDoing`, `RSCG_WhatIAmDoing_Common`, and `rscgutils` packages in various `.csproj` files. Added a new namespace `WhatIAmDoingData` to `Program.cs` and commented out arguments for the `generateFiles` command in the same file. Moved verbose logging to the `DisplayData` class and added a new `Verbose` property to control it. Added a new method `ListSDKS` to the `ProcessOutput` class and used it in `GenerateFiles.cs`. Added `InterceptInstanceClass` attribute to `InterceptorMethodInstanceClass` in `Interceptors.cs`.